Move recycler view adapters to fragments
