+++ app/src/main/java/com/example/anywherefitness/ui/Instructor/InstructorClassesViewModel.kt
@@ -1,15 +1,9 @@
 package com.example.anywherefitness.ui.Instructor
 
 import android.app.Application
-import android.content.Context
-import android.view.View
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.example.anywherefitness.FitnessClassAdapter
 import com.example.anywherefitness.database.UserRepo
 import com.example.anywherefitness.model.FitnessClass
 
@@ -39,17 +33,4 @@ class InstructorClassesViewModel(application: Application) : AndroidViewModel(ap
         var instructorFitnessClassList: MutableList<FitnessClass> = repo.getAllFitnessClasss()
         return instructorFitnessClassList
     }
-
-    fun setupRecycler(context: Context?,
-                      fitnessClassList: MutableList<FitnessClass>,
-                      recyclerView: RecyclerView,
-                      repo: UserRepo) {
-        recyclerView.apply {
-            setHasFixedSize(false)
-            layoutManager = LinearLayoutManager(context)
-            adapter = FitnessClassAdapter(fitnessClassList, repo)
-        }
-    }
-
-
 }
